Crate parsimonious

Parsimonious: a parser combinator library for Rust

The goal of this library is to provide parser combinators that:

are optimized for LL(1) grammars,

support streaming input,

do as little buffering or copying as possible, and

do as little dynamic method dispatch as possible.

It is based on "Monadic Parsing in Haskell" by Hutton and Meijer, JFP 8(4) pp. 437-444.

Structs

AndThenGuardedParser

BoxableParser

BufferedGuardedParser

BufferedStatefulParser

CharacterParser

CharacterStatefulParser

ImpossibleStatefulParser

MapGuardedParser

MapStatefulParser

OrElseGuardedParser

OrEmitParser

PlusParser

StarParser

StarStatefulParser

Enums

AndThenStatefulParser

GuardedParseResult

Impossible

OrElseStatefulParser

OrEmitStatefulParser

ParseResult

Str

Traits

BoxableParserOf

Consumer
A trait for consumers of data, typically buffers.

GuardedParser

GuardedParserOf

Parser

ParserOf

StatefulParserOf
A trait for stateful parsers.

Functions

character

ignore