Crate parsimonious

Parsimonious: a parser combinator library for Rust

The goal of this library is to provide parser combinators that:

are optimized for LL(1) grammars,

support streaming input,

do as little buffering or copying as possible, and

do as little dynamic method dispatch as possible.

It is based on:

Monadic Parsing in Haskell by G. Hutton and E. Meijer, JFP 8(4) pp. 437-444,

Nom, eating data byte by byte by G. Couprie.

Repo | Crate | CI

Modules

impls
Provide implementations of parser traits.

Enums

GuardedParseResult
The result of a guarded parse.

ParseResult
The result of a parse.

Str

Traits

BoxableParserOf
A trait for boxable parsers.

Consumer
A trait for consumers of data, typically buffers.

GuardedParserOf
A trait for stateless guarded parsers.

ParserOf
A trait for stateless parsers.

StatefulParserOf
A trait for stateful parsers.

Functions

character

character_guard

ignore